+++ src/core/decorators.py
@@ -9,7 +9,7 @@ from pydantic import ValidationError
 
 # Local application imports
 from core.messager.messager import Messager
-from core.agent import Agent  # Keep if Agent injection is needed
+from core.agent.agent import Agent  # Keep if Agent injection is needed
 
 # Assuming from_mqtt_message is updated or replaced to handle aiomqtt.Message
 from core.protocol.message import from_payload, AnyMessage, BaseMessage
