+++ src/core/messager/drivers/RedisDriver.py
@@ -1,10 +1,8 @@
-from core.messager.drivers import BaseDriver, DriverConfig
+from core.messager.drivers import BaseDriver, DriverConfig, MessageHandler
 from core.protocol.message import BaseMessage, from_payload
 
-
 import asyncio
-import json
-from typing import Any, Callable, Coroutine, Optional
+from typing import Optional
 import aioredis
 
 
@@ -24,15 +22,13 @@ class RedisDriver(BaseDriver):
         if self._redis:
             await self._redis.close()
 
-    async def publish(self, topic: str, payload: Any, **kwargs) -> None:
+    async def publish(self, topic: str, payload: BaseMessage, **kwargs) -> None:
         # Handle BaseMessage serialization
-        if isinstance(payload, BaseMessage):
-            data = payload.model_dump_json()
-        else:
-            data = payload if isinstance(payload, (bytes, str)) else json.dumps(payload)
+        data = payload.model_dump_json()
+
         await self._redis.publish(topic, data)
 
-    async def subscribe(self, topic: str, handler: Callable[[Any], Coroutine], **kwargs) -> None:
+    async def subscribe(self, topic: str, handler: MessageHandler, **kwargs) -> None:
         pubsub = self._redis.pubsub()
         await pubsub.subscribe(topic)
