+++ src/main.py
@@ -7,7 +7,7 @@ from dotenv import load_dotenv
 
 # Core components (assuming they are updated)
 from core.messager.messager import Messager
-from core.agent import Agent
+from core.agent.agent import Agent
 from core.tool_manager import ToolManager
 from core.logger import get_logger, LogLevel, parse_log_level
 from core.protocol.message import (
